Fixed DTO naming conventions and fixed folder structure
