--- src/main/java/com/example/dashboardbackend/dtos/ChangeUserRoleRequest.java
@@ -1,9 +1,0 @@
-package com.example.dashboardbackend.dtos;
-
-import com.example.dashboardbackend.models.enums.UserRole;
-
-public record ChangeUserRoleRequest(
-        UserRole userRole
-) {
-
-}
