--- src/main/java/com/example/dashboardbackend/dtos/CreateFamilyRequest.java
@@ -1,8 +1,0 @@
-package com.example.dashboardbackend.dtos;
-
-public record CreateFamilyRequest(
-        String familyName,
-        String password,
-        String email
-) {
-}
