--- src/main/java/com/example/dashboardbackend/dtos/FamilyResponse.java
@@ -1,8 +1,0 @@
-package com.example.dashboardbackend.dtos;
-
-public record FamilyResponse(
-        Long id,
-        String familyName,
-        String email
-) {
-}
