--- src/main/java/com/example/dashboardbackend/dtos/LoginRequest.java
@@ -1,7 +1,0 @@
-package com.example.dashboardbackend.dtos;
-
-public record LoginRequest(
-        String name,
-        String password
-) {
-}
