--- src/main/java/com/example/dashboardbackend/dtos/MediaUploadRequest.java
@@ -1,8 +1,0 @@
-package com.example.dashboardbackend.dtos;
-
-import org.springframework.web.multipart.MultipartFile;
-
-public record MediaUploadRequest(
-        MultipartFile multipartFile
-) {
-}
