--- src/main/java/com/example/dashboardbackend/dtos/StatusResponse.java
@@ -1,6 +1,0 @@
-package com.example.dashboardbackend.dtos;
-
-public record StatusResponse(
-        String role
-) {
-}
