--- src/main/java/com/example/dashboardbackend/dtos/SysAdminLoginResponse.java
@@ -1,6 +1,0 @@
-package com.example.dashboardbackend.dtos;
-
-public record SysAdminLoginResponse(
-        String type
-) {
-}
