--- src/main/java/com/example/dashboardbackend/dtos/UpdateFamilyNameRequest.java
@@ -1,6 +1,0 @@
-package com.example.dashboardbackend.dtos;
-
-public record UpdateFamilyNameRequest(
-        String familyName
-) {
-}
