--- src/main/java/com/example/dashboardbackend/dtos/UserSelectRequest.java
@@ -1,7 +1,0 @@
-package com.example.dashboardbackend.dtos;
-
-public record UserSelectRequest(
-        Long userId,
-        String pin
-) {
-}
