+++ src/main/java/com/example/dashboardbackend/dtos/auth/LoginRequest.java
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2026 SAP SE or an SAP affiliate company. All rights reserved.
+ */
+
+package com.example.dashboardbackend.dtos.auth;
+
+public record LoginRequest(
+        String name,
+        String password
+) {
+}
