+++ src/main/java/com/example/dashboardbackend/dtos/dashboard/StatusResponse.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright (c) 2026 SAP SE or an SAP affiliate company. All rights reserved.
+ */
+
+package com.example.dashboardbackend.dtos.dashboard;
+
+public record StatusResponse(
+        String role
+) {
+}
