+++ src/main/java/com/example/dashboardbackend/dtos/family/CreateFamilyRequest.java
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2026 SAP SE or an SAP affiliate company. All rights reserved.
+ */
+
+package com.example.dashboardbackend.dtos.family;
+
+public record CreateFamilyRequest(
+        String familyName,
+        String password,
+        String email
+) {
+}
