+++ src/main/java/com/example/dashboardbackend/dtos/family/FamilyLoginResponse.java
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2026 SAP SE or an SAP affiliate company. All rights reserved.
+ */
+
+package com.example.dashboardbackend.dtos.family;
+
+import com.example.dashboardbackend.dtos.user.UserProfileResponse;
+
+import java.util.List;
+
+public record FamilyLoginResponse(
+        Long familyId,
+        List<UserProfileResponse> profiles,
+        String role
+) {
+}
