+++ src/main/java/com/example/dashboardbackend/dtos/user/ChangeUserRoleRequest.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2026 SAP SE or an SAP affiliate company. All rights reserved.
+ */
+
+package com.example.dashboardbackend.dtos.user;
+
+import com.example.dashboardbackend.models.enums.UserRole;
+
+public record ChangeUserRoleRequest(
+        UserRole userRole
+) {
+
+}
