+++ src/main/java/com/example/dashboardbackend/dtos/user/UserProfileResponse.java
@@ -1,9 +1,13 @@
-package com.example.dashboardbackend.dtos;
+/*
+ * Copyright (c) 2026 SAP SE or an SAP affiliate company. All rights reserved.
+ */
+
+package com.example.dashboardbackend.dtos.user;
 
 import com.example.dashboardbackend.models.enums.UserAvatarType;
 import com.example.dashboardbackend.models.enums.UserRole;
 
-public record UserProfile(
+public record UserProfileResponse(
         Long id,
         String name,
         String avatar,
