+++ src/main/java/com/example/dashboardbackend/dtos/user/UserSelectRequest.java
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2026 SAP SE or an SAP affiliate company. All rights reserved.
+ */
+
+package com.example.dashboardbackend.dtos.user;
+
+public record UserSelectRequest(
+        Long userId,
+        String pin
+) {
+}
