+++ src/main/java/com/example/dashboardbackend/dtos/weather/GeoLocationResponse.java
@@ -2,7 +2,12 @@
 
 package com.example.dashboardbackend.dtos.weather;
 
-import java.util.List;
-
-public record GeoLocationResponse(List<GeoLocation> results) {
+public record GeoLocationResponse(
+        String name,
+        double latitude,
+        double longitude,
+        String country,
+        String timezone,
+        String admin1 // Bundesland in open meteo
+) {
 }
