+++ src/main/java/com/example/dashboardbackend/dtos/weather/WeatherResponse.java
@@ -3,7 +3,7 @@ package com.example.dashboardbackend.dtos.weather;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record WeatherResponse(
-    @JsonProperty("current") Current current,
-    @JsonProperty("daily") Daily daily
+    @JsonProperty("current") CurrentResponse currentResponse,
+    @JsonProperty("daily") DailyResponse dailyResponse
 ) {
 }
