+++ src/main/java/com/example/dashboardbackend/dtos/widgets/UpdateWidgetConfigRequest.java
@@ -1,4 +1,8 @@
-package com.example.dashboardbackend.dtos;
+/*
+ * Copyright (c) 2026 SAP SE or an SAP affiliate company. All rights reserved.
+ */
+
+package com.example.dashboardbackend.dtos.widgets;
 
 import com.example.dashboardbackend.models.widgets.WidgetConfig;
